SecurityRuntime, Windows/File based credential managers work
Removed "FileEncryption" namespace, moved things up one.

The credential manager exposed in the SecurityRuntime (and to be used by
the app) is actually a CredentialHub on top of a file based credential
manager that is encrypted via AES256 using an encryption key that is
stored in the platform's credential manager (such as the Android
KeyStore, iOS KeyChain or Windows Credential Manager). This helps
getting around any size restrictions on the platform's credential
manager).